code cleaning and rearrangement (#16)
* review corrections/suggestions 1
* code cleaning and rearrangement
* address review comments
