+++ library/src/include/lib_host_helpers.hpp
@@ -92,6 +92,74 @@ I get_index(I* intervals, I max, I dim)
     return i;
 }
 
+template <typename T, typename I>
+static void call_swap(I& n, T& x_in, I& incx, T& y_in, I& incy)
+{
+    T* const x = &(x_in);
+    T* const y = &(y_in);
+    for(I i = 0; i < n; i++)
+    {
+        I const ix = i * incx;
+        I const iy = i * incy;
+
+        T const temp = x[ix];
+        x[ix] = y[iy];
+        y[iy] = temp;
+    }
+}
+
+static float real_part(float z)
+{
+    return (z);
+};
+static float real_part(std::complex<float> z)
+{
+    return (z.real());
+};
+static float real_part(rocblas_complex_num<float> z)
+{
+    return (z.real());
+};
+
+static double real_part(double z)
+{
+    return (z);
+};
+static double real_part(std::complex<double> z)
+{
+    return (z.real());
+};
+static double real_part(rocblas_complex_num<double> z)
+{
+    return (z.real());
+};
+
+static float imag_part(float z)
+{
+    return (0);
+};
+static float imag_part(std::complex<float> z)
+{
+    return (z.imag());
+};
+static float imag_part(rocblas_complex_num<float> z)
+{
+    return (z.imag());
+};
+
+static double imag_part(double z)
+{
+    return (0);
+};
+static double imag_part(std::complex<double> z)
+{
+    return (z.imag());
+};
+static double imag_part(rocblas_complex_num<double> z)
+{
+    return (z.imag());
+};
+
 #ifdef ROCSOLVER_VERIFY_ASSUMPTIONS
 // Ensure __assert_fail is declared.
 #if !__is_identifier(__assert_fail)
@@ -151,6 +219,14 @@ extern "C" [[noreturn]] void __assert_fail(const char* assertion,
     } while(0)
 #else
 #define ROCSOLVER_ASSUME_X(invariant, msg) __builtin_assume(invariant)
+#endif
+
+#ifndef CHECK_HIP
+#define CHECK_HIP(fcn)                  \
+    {                                   \
+        hipError_t const istat = (fcn); \
+        assert(istat == hipSuccess);    \
+    }
 #endif
 
 ROCSOLVER_END_NAMESPACE
